requirements.txt
pandas
Flask
Flask-Cors
Flask-RESTful
python-dotenv
mysql-connector-python
